rename old combobox as LegacyComboBox
